+++ src/popup/utils/messenger.ts
@@ -16,10 +16,8 @@
 
 import { v4 } from 'uuid';
 import { browser } from 'webextension-polyfill-ts';
-import { Runtime } from 'webextension-polyfill-ts/dist/generated/runtime';
-import { POPUP_PORT_NAME } from '../../config';
 import { MessageOrigin } from '../../types';
-import { RuntimePortDuplex } from '../../utils/RuntimePortDuplex';
+import { RuntimePortDuplex } from '../../streamUtils/RuntimePortDuplex';
 
 export function initConnection(origin: MessageOrigin.TOOLBAR | MessageOrigin.SIGN_POPUP) {
   try {
@@ -31,34 +29,3 @@ export function initConnection(origin: MessageOrigin.TOOLBAR | MessageOrigin.SIG
   }
 }
 
-// type Callback = (message: any, port: Runtime.Port) => void;
-//
-// class Messenger {
-//   port: Runtime.Port;
-//
-//   constructor(name: string) {
-//     try {
-//       this.port = browser.runtime.connect(null, { name });
-//     } catch (error) {
-//       // fails when running as webpage
-//     }
-//   }
-//
-//   addListener = (callback: Callback) => {
-//     this.port.onMessage.addListener(callback);
-//   };
-//
-//   removeListener = (callback: Callback) => {
-//     this.port.onMessage.removeListener(callback);
-//   };
-//
-//   send = (data: any) => {
-//     try {
-//       this.port.postMessage(data);
-//     } catch (error) {
-//       // fails when running as webpage
-//     }
-//   };
-// }
-//
-// export default new Messenger(POPUP_PORT_NAME);
